split off configuration into a service to load IndexMetadata; processed comments of Domi
